Add base class for providers
Removed some duplicated configuration using the base class
